requirements.txt
torch>=2.0.0
numpy>=1.24.0
cryptography>=41.0.0
pyjwt>=2.8.0
